Rename data folder and update paths
