+++ README.md
@@ -17,9 +17,9 @@ produce audio using ElevenLabs.
 
 ## Data files
 
-- `verbs.csv` – list of verbs and common collocations.
-- `forms.csv` – grammatical forms with trigger phrases.
-- `persons.csv` – grammatical persons.
+- `verb_data/verbs.csv` – list of verbs and common collocations.
+- `verb_data/tenses.csv` – grammatical tenses with trigger phrases.
+- `verb_data/persons.csv` – grammatical persons.
 - `cards.csv` – main table used for each flashcard; generated by
   `generate_cards_init.py` and later filled out by other tools.
 
@@ -31,12 +31,12 @@ produce audio using ElevenLabs.
   sentences using the OpenAI API.
 - `generate_audio.py` creates audio files for the sentences with
   ElevenLabs.
-- `get_conjugation_rae.py` scrapes conjugations from the RAE dictionary
-  website and outputs them in the project's JSON format. Reflexive verbs
-  are handled automatically by removing the trailing pronoun before
-  fetching.
-- `regular_form_generator.py` provides hypothetical regular forms used to
-  flag irregular conjugations.
+- `utilities/get_conjugation_rae.py` scrapes conjugations from the RAE
+  dictionary website and outputs them in the project's JSON format.
+  Reflexive verbs are handled automatically by removing the trailing
+  pronoun before fetching.
+- `utilities/regular_form_generator.py` provides hypothetical regular
+  forms used to flag irregular conjugations.
 
 ## `cards.csv` fields
