+++ tests/test_get_conjugation_rae_ar_reflexives.py
@@ -3,8 +3,8 @@ import sys
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 
-from get_conjugation_rae import RAEConjugationFetcher
-from get_conjugation_rae import RAEConjugationTransformer
+from utilities.get_conjugation_rae import RAEConjugationFetcher
+from utilities.get_conjugation_rae import RAEConjugationTransformer
 
 
 def test_darse():
