+++ tests/test_get_conjugation_rae_non_ar_reflexives.py
@@ -3,9 +3,9 @@ import sys
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 
-from get_conjugation_rae import RAEConjugationFetcher
-from get_conjugation_rae import RAEConjugationTransformer
-from get_conjugation_rae import strip_reflexive
+from utilities.get_conjugation_rae import RAEConjugationFetcher
+from utilities.get_conjugation_rae import RAEConjugationTransformer
+from utilities.get_conjugation_rae import strip_reflexive
 
 
 def test_meterse():
